plug-ins: Mitigate ZDI-CAN-29396
Resolves #16249
This patch adds a check for the SGI read in
bpp being within acceptable ranges (1 or 2).
It also casts pointer arithmetic to gsize
(similar to other areas of the SGI code)
to prevent overflow.

(cherry picked from commit 76531da9732f38566e5fd8f8f80c837158511ae5)

+++ plug-ins/file-sgi/sgi.c
@@ -474,7 +474,7 @@ load_image (GFile   *file,
   pixels[0]   = g_new (guchar, ((gsize) tile_height) * sgip->xsize * sgip->bpp * bytes);
 
   for (i = 1; i < tile_height; i ++)
-    pixels[i] = pixels[0] + sgip->xsize * sgip->bpp * bytes * i;
+    pixels[i] = pixels[0] + (((gsize) sgip->xsize) * sgip->bpp * bytes * i);
 
   rows    = g_new (unsigned short *, sgip->zsize);
   rows[0] = g_new (unsigned short, ((gsize) sgip->xsize) * sgip->zsize);
